python server : Separate class per file - load database into web.py and share it - web_server.sh now can pass multiple parameter to __main__.py
