+++ src/web/db.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 import sys
 from tinydb import TinyDB, Query
 import uuid
