+++ web_server.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-# Launch python server on http port 8000
-cd src/web; python3 __main__.py
+# Launch python server on http port 8000 by default
+cd src/web; python3 __main__.py "$@"
